+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 # CHANGELOG
 
 <p>
+<b>Version 2.1.52</b> - October 2023<br/>
+- HUE Light node: fixed another possible switch on brightness issue.<br/>
+</p>
+<p>
 <b>Version 2.1.51</b> - October 2023<br/>
 - HUE Light node: fixed a possible switch on brightness issue.<br/>
 </p>